HUGE refactoring to rename files and class names to work well with upcoming changes
